+++ rust/stackable-cockpit/src/platform/manifests.rs
@@ -22,7 +22,7 @@ use crate::{
 #[derive(Debug, Snafu)]
 pub enum Error {
     /// This error indicates that parsing a string into a path or URL failed.
-    #[snafu(display("failed to parse '{path_or_url}' as path/url"))]
+    #[snafu(display("failed to parse {path_or_url:?} as path/url"))]
     ParsePathOrUrl {
         source: PathOrUrlParseError,
         path_or_url: String,