+++ rust/stackable-cockpit/src/utils/params.rs
@@ -34,7 +34,7 @@ pub enum IntoParametersError {
     #[snafu(display("failed to parse raw parameter"))]
     RawParse { source: RawParameterParseError },
 
-    #[snafu(display("invalid parameter '{parameter}', expected one of {expected}"))]
+    #[snafu(display("invalid parameter {parameter:?}, expected one of {expected:?}"))]
     InvalidParameter { parameter: String, expected: String },
 }
 
@@ -106,7 +106,7 @@ pub enum RawParameterParseError {
 
 impl Display for RawParameter {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}={}", self.name, self.value)
+        write!(f, "{name}={value}", name = self.name, value = self.value)
     }
 }