+++ rust/stackablectl/CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
   This should unblock demos to run in all namespaces, as they can template the namespace e.g. for the FQDN of services ([#355]).
 - Add progress reporting ([#376]).
 - Include SparkConnectServer in the `stacklet list` output ([#380]).
+- Add release upgrade functionality to `stackablectl release` command through `upgrade` subcommand ([#379]).
 
 ### Changed
 
@@ -24,6 +25,7 @@ All notable changes to this project will be documented in this file.
 [#373]: https://github.com/stackabletech/stackable-cockpit/pull/373
 [#376]: https://github.com/stackabletech/stackable-cockpit/pull/376
 [#380]: https://github.com/stackabletech/stackable-cockpit/pull/380
+[#379]: https://github.com/stackabletech/stackable-cockpit/pull/379
 
 ## [25.3.0] - 2025-03-27